+++ app/panel/account/page.tsx
@@ -7,27 +7,7 @@ import Link from "next/link";
 import { Suspense } from "react";
 import { getUserProfile, UserProfile } from "../profile/profile";
 import OrgsList from "../organizationsList";
-import FormButton from "../../(components)/FormButton";
-
-export async function logout() {
-    const token = cookies().get("token") || redirect("/account/login");
-
-    const result = await fetch(
-        `${process.env.API_URL}/account/logout`,
-        {
-            method: "GET",
-            headers: {
-                Authorization: token.value,
-            },
-        }
-    );
-
-    if (result.status === 200) {
-        cookies().delete('token');
-    }
-
-    redirect("/");
-}
+import Logout from "../../(components)/Logout";
 
 export default async function Page() {
     const cookie = cookies();
@@ -68,12 +48,7 @@ export default async function Page() {
                                 EDIT
                             </Button>
                         </Link>
-                        <FormButton onClick={async () => {
-                            "use server"
-                            await logout();
-                        }}>
-                            <Button variant="tonal" size="small">Logout</Button>
-                        </FormButton>
+                        <Logout />
                     </Row>
                 </Stack>
             </Row>
