+++ app/panel/organization/edit/[orgId]/page.tsx
@@ -1,4 +1,4 @@
-import { Button, Container, InputField, Row, Spacer, Stack } from "@yakad/ui";
+import { Button, Container, InputField, Row, Stack } from "@yakad/ui";
 import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
 import { Organization } from "../../../organization";
